Refactored completely... only a single run through the grid is necessary for both parts
